// frontend/src/api/generated/zod/user-profile/user-profile.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import { z as zod } from "zod";

export const updateUserProfileBody = zod.object({
  firstName: zod.string(),
  lastName: zod.string(),
  email: zod.string(),
  phoneNumber: zod.string(),
  dateFormat: zod.string().optional(),
});

export const changePasswordBodyNewPasswordMin = 8;

export const changePasswordBodyNewPasswordMax = 2147483647;

export const changePasswordBody = zod.object({
  currentPassword: zod.string(),
  newPassword: zod
    .string()
    .min(changePasswordBodyNewPasswordMin)
    .max(changePasswordBodyNewPasswordMax),
  confirmationPassword: zod.string(),
});